Changed data criteria filtering and measure test corrections
Data filtering was removing elements that should have been kept,
this is now closer to what it should be taking out and is much
easier to change. The test has been corrected to handle denominator
exclusions of any name that are unnecessary, and fixed a bug with
the individual measure correction CMS82v3.