Removed subset inclusive, added additional test exclusions
Removed (for now) subset inclusive operator, since it was
conflicting with the success of other measures.  This is to be
investigated in the near future, but, according to the comments
only affects 2 measures, so this will be worked out at a later
point.

Also added two measure specific exclusions to the test.