Cleaned up tester, corrected 3 sets of errors
Cleaned up the tester structure to be better organized. Added an
additional inclusive test to types to correct cases where the
criteria should have been inclusive but wasn't showing that.
Added code to limit copying of specific occurence data into the
source referring to it, and erased line that created verbose
references.